separate theorydb into nnpdf_data
