+++ doc/sphinx/make_theory_csv.py
@@ -6,7 +6,7 @@ from argparse import ArgumentParser
 from pathlib import Path
 
 from nnpdf_data import theory_cards
-from validphys.theorydbutils import fetch_all
+from nnpdf_data.theorydbutils import fetch_all
 
 if __name__ == "__main__":
     parser = ArgumentParser()
