+++ validphys2/src/validphys/utils.py
@@ -7,9 +7,6 @@ from typing import Any, Hashable, Mapping, Sequence
 
 from frozendict import frozendict
 import numpy as np
-from validobj import ValidationError, parse_input
-
-from reportengine.compat import yaml
 
 
 def make_hashable(obj: Any):
@@ -46,54 +43,6 @@ def generate_path_filtered_data(fit_path, setname):
     data_path = fit_path / "filter" / setname / f"filtered_data_{setname}.yaml"
     unc_path = data_path.with_name(f"filtered_uncertainties_{setname}.yaml")
     return data_path, unc_path
-
-
-def parse_yaml_inp(input_yaml, spec):
-    """
-    Helper function to parse yaml using the `validobj` library and print
-    useful error messages in case of a parsing error.
-
-    https://validobj.readthedocs.io/en/latest/examples.html#yaml-line-numbers
-    """
-    input_yaml = pathlib.Path(input_yaml)
-    inp = yaml.round_trip_load(input_yaml.open("r", encoding="utf-8"))
-    try:
-        return parse_input(inp, spec)
-    except ValidationError as e:
-        current_exc = e
-        current_inp = inp
-        error_text_lines = []
-        while current_exc:
-            if hasattr(current_exc, 'wrong_field'):
-                wrong_field = current_exc.wrong_field
-                # Mappings compping from ``round_trip_load`` have an
-                # ``lc`` attribute that gives a tuple of
-                # ``(line_number, column)`` for a given item in
-                # the mapping.
-                line = current_inp.lc.item(wrong_field)[0]
-                error_text_lines.append(f"Problem processing key at line {line} in {input_yaml}:")
-                current_inp = current_inp[wrong_field]
-            elif hasattr(current_exc, 'wrong_index'):
-                wrong_index = current_exc.wrong_index
-                # Similarly lists allow to retrieve the line number for
-                # a given item.
-                line = current_inp.lc.item(wrong_index)[0]
-                current_inp = current_inp[wrong_index]
-                error_text_lines.append(
-                    f"Problem processing list item at line {line} in {input_yaml}:"
-                )
-            elif hasattr(current_exc, 'unknown'):
-                unknown_lines = []
-                for u in current_exc.unknown:
-                    unknown_lines.append((current_inp.lc.item(u)[0], u))
-                unknown_lines.sort()
-                for line, key in unknown_lines:
-                    error_text_lines.append(
-                        f"Unknown key {key!r} defined at line {line} in {input_yaml}:"
-                    )
-            error_text_lines.append(str(current_exc))
-            current_exc = current_exc.__cause__
-        raise ValidationError('\n'.join(error_text_lines)) from e
 
 
 @contextlib.contextmanager
